add junit tests (add and remove developer, getCommentsforTask, set and get Dependencies) and fixed wrong code, added one task_dependency into dummydata

+++ src/main/java/de/htwsaar/owlkeeper/storage/model/TaskModel.java
@@ -81,7 +81,7 @@ public class TaskModel extends AbstractModel<Task, TaskDao> {
      * @return dependingTask
      */
 
-    public List<Integer> getDependencies () {
+    public List<Long> getDependencies () {
         long taskId = this.getContainer().getId();
         return DBConnection.getJdbi().withExtension(TaskDao.class, (dao -> dao.getDependencies(taskId)));
     }
